+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 7.2.0
 
-* default format24Hours = true in timePicker
+* change format24Hours -> format12Hours in timePicker
 
 ## 7.1.0